chore : renaming typescript folder into core
