+++ src/components/Avatar/core/avatar.stories.tsx
