+++ src/components/Avatar/core/avatar.styles.ts
