+++ src/components/Avatar/core/avatar.types.ts
