+++ src/shared/icons/IconSearch/core/iconSearch.ts
